principalEmzima/pythonAutomatizacion/SaucedemoProjectPlaySync/imports.py
import pytest
import allure
import time
import os
import subprocess

from pytest_bdd import scenario, given, when,then,parsers
from playwright.sync_api import Page,sync_playwright,Playwright,expect
from typing import Literal

