principalEmzima/pythonAutomatizacion/SaucedemoProjectPlaySync/requirements.txt
allure-pytest==2.13.5
behave==1.2.6
pytest==8.3.3
pytest-html==4.1.1
selenium==4.25.0
webdriver-manager==4.0.1
pytest-bdd==7.3.0 
python==3.12.6